Add support for TMY3 weather files
Changed weather and sun classes to handle TMY3. As solar irradiation is
not given for a horizontal plane, the thetaZ angle has to be computed
first, to transform the data to the TRY standard.

Also updated the weather test file to include an example for TMY3 data.

Further, a test file for New York's JFK international airport is included.